Bug fixes
-Custom validations
-Code cleaup
-bug fixes
