+++ suppliers/src/UI/embc-supplier/src/app/review/review.component.html
@@ -38,7 +38,7 @@
                 <p>{{supplier?.remittanceAddress?.city}}</p>
                 <p>{{supplier?.remittanceAddress?.province}}</p>
                 <p>{{supplier?.remittanceAddress?.postalCode}}</p>
-                <p>{{supplier?.businessCountry.name}}</p>
+                <p>{{supplier?.businessCountry?.name}}</p>
             </div>
             <div *ngIf="supplierSubmissionType === 'invoice'">
                 <div class="formHeading">
